refactor: move edit actions to own folder
