+++ src/components/message/edit-message-actions/styles.scss
@@ -1,0 +1,17 @@
+@use '~@zero-tech/zui/styles/theme' as theme;
+
+.edit-message-actions {
+  display: flex;
+  align-items: center;
+  align-self: center;
+  gap: 8px;
+
+  &__icon {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 32px;
+    width: 32px;
+    color: theme.$color-greyscale-12;
+  }
+}
